clausewise/requirements.txt
streamlit
python-docx
PyPDF2
python-dotenv
ibm-watson
transformers
sentencepiece
streamlit-tags